+++ lib/vector-core/src/lib.rs
@@ -25,6 +25,7 @@
 pub mod api;
 pub mod config;
 pub mod event;
+pub mod fanout;
 pub mod mapping;
 pub mod metrics;
 pub mod sink;
